Move getApiHelper to api.ts and getDatadogHost to util.ts
